+++ src/components/Sidebar/RvSidebarHeader.tsx
@@ -2,34 +2,41 @@
 
 import { useMemo } from "react";
 import type { FC } from "react";
-import { RvFlex } from "../Box";
+import { RvHStack } from "../Box";
 import type { RvSidebarHeaderFCType, StyledSidebarHeaderType } from "../../types";
 
-const rvSidebarHeaderDefaultProps: StyledSidebarHeaderType = {
-	width: "",
+const defaultProps: StyledSidebarHeaderType = {
+	width: "100%",
 	height: "50px",
+	borderRadius: 0,
 	alignItems: "center",
 	justifyContent: "center",
 	textColor: "black300",
 	bgColor: "black600",
+	hover: {
+		bgColor: "none",
+		textColor: "none",
+	},
 };
 
 const RvSidebarHeader: FC<RvSidebarHeaderFCType> = ({ children, ...props }) => {
 	const {
-		width = props.width ?? rvSidebarHeaderDefaultProps.width,
-		height = props.height ?? rvSidebarHeaderDefaultProps.height,
-		alignItems = props.alignItems ?? rvSidebarHeaderDefaultProps.alignItems,
-		justifyContent = props.justifyContent ?? rvSidebarHeaderDefaultProps.justifyContent,
-		textColor = props.textColor ?? rvSidebarHeaderDefaultProps.textColor,
-		bgColor = props.bgColor ?? rvSidebarHeaderDefaultProps.bgColor,
+		width = props.width ?? defaultProps.width,
+		height = props.height ?? defaultProps.height,
+		borderRadius = props.borderRadius ?? defaultProps.borderRadius,
+		alignItems = props.alignItems ?? defaultProps.alignItems,
+		justifyContent = props.justifyContent ?? defaultProps.justifyContent,
+		textColor = props.textColor ?? defaultProps.textColor,
+		bgColor = props.bgColor ?? defaultProps.bgColor,
+		hover = props.hover ?? defaultProps.hover,
 	} = useMemo(() => props, [props]);
 
 	const newProps = useMemo(
-		() => ({ ...props, width, height, alignItems, justifyContent, bgColor, textColor }),
-		[props, width, height, alignItems, justifyContent, bgColor, textColor]
+		() => ({ ...props, width, height, borderRadius, alignItems, justifyContent, bgColor, textColor, hover }),
+		[props, width, height, borderRadius, alignItems, justifyContent, bgColor, textColor, hover]
 	);
 
-	return useMemo(() => <RvFlex {...newProps}>{children}</RvFlex>, [newProps]);
+	return useMemo(() => <RvHStack {...newProps}>{children}</RvHStack>, [newProps]);
 };
 
 export default RvSidebarHeader;