+++ src/components/Sidebar/RvSubMenu.tsx
@@ -6,7 +6,7 @@ import { SubMenu } from "react-pro-sidebar";
 import type { SubMenuProps } from "react-pro-sidebar";
 import type { RvSubMenuFCType } from "../../types/components/Sidebar";
 
-export const rvSubMenuDefaultProps: SubMenuProps = {};
+export const defaultProps: SubMenuProps = {};
 
 const RvSubMenu: FC<RvSubMenuFCType> = ({ children, ...props }) => {
 	const newProps = useMemo(() => ({ ...props }), [props]);